Extract ImapFolder from ImapStore
